#33, #34 Обновил структуру приложения. Вынес файлы моделей в отдельную папку. Удалил models.py из приложения.
