+++ test_bim/models/answer_bim.py
@@ -1,0 +1,24 @@
+from django.db import models
+from modules.models.modules import AbstractModel
+
+
+# Create your models here.
+
+
+class AnswerBim(AbstractModel):
+    answer = models.CharField(max_length=100, verbose_name='Ответ')
+    is_correct = models.BooleanField(verbose_name='Валидация ответа',
+                                     choices=[(False, 'Неверный ответ'), (True, 'Верный ответ')],
+                                     default=False)
+    question_bim = models.ForeignKey('test_bim.QuestionBim',
+                                     related_name='answer_bim',
+                                     on_delete=models.CASCADE,
+                                     verbose_name='Вопрос')
+
+    def __str__(self):
+        return f'{self.answer}'
+
+    class Meta:
+        db_table = 'answer_bim'
+        verbose_name = 'Ответ Bim'
+        verbose_name_plural = 'Ответ Bim'
